move media query debugger into src/components
